html/editing/focus/focus-02-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focus - key events</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/editing.html#focus">
<meta assert="flag" content="interact">
<meta assert="assert" content="Check if the key events recieved by document are targeted at the body element when no element is focused">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h2>Steps:</h2>
<ol>
  <li>Press any key in [0-9a-zA-Z].</li>
</ol>
<h2>Expected results:</h2>
<p>PASS</p>
<div id="log"></div>
<script>

var t1 = async_test("The keydown event must be targeted at the body element"),
    t2 = async_test("The keypress event must be targeted at the body element"),
    t3 = async_test("The keyup event must be targeted at the body element");

setup({timeout: 10000});

document.addEventListener("keydown", function(evt){
  t1.step(function() {
    assert_equals(evt.target, document.body, "The keydown events must be targeted at the body element.");
  });
  t1.done();
}, false);

document.addEventListener("keypress", function(evt){
  t2.step(function() {
    assert_equals(evt.target, document.body, "The keypress events must be targeted at the body element.");
  });
  t2.done();
}, false);

document.addEventListener("keyup", function(evt){
  t3.step(function() {
    assert_equals(evt.target, document.body, "The keyup event must be targeted at the body element.");
  });
  t3.done()
}, false);

</script>
